+++ src/bam/collapse_reads_to_fragments/pipeline.rs
@@ -1,4 +1,5 @@
 use super::calc_consensus::{CalcNonOverlappingConsensus, CalcOverlappingConsensus};
+use super::unmark_record;
 use anyhow::Result;
 use bio::io::fastq;
 use derive_new::new;
@@ -9,6 +10,7 @@ use std::cmp::Ordering;
 use std::collections::{BTreeMap, HashMap, HashSet};
 use std::io;
 use std::ops::Deref;
+use std::ops::DerefMut;
 use uuid::Uuid;
 
 #[derive(new)]
@@ -18,7 +20,7 @@ pub struct CallConsensusRead<W: io::Write> {
     fq2_writer: fastq::Writer<W>,
     fq_se_writer: fastq::Writer<W>,
     bam_skipped_writer: bam::Writer,
-    verbose_read_names: bool,
+    annotate_record_ids: bool,
 }
 
 type Position = i64;
@@ -102,7 +104,7 @@ impl<W: io::Write> CallConsensusRead<W> {
         let mut duplicate_groups: HashMap<GroupId, RecordIDs> = HashMap::new();
         let mut record_storage: HashMap<RecordId, RecordStorage> = HashMap::new();
         let mut current_chrom = None;
-        let mut read_ids: Option<HashMap<usize, Vec<u8>>> = if self.verbose_read_names {
+        let mut read_ids: Option<HashMap<usize, Vec<u8>>> = if self.annotate_record_ids {
             Some(HashMap::new())
         } else {
             None
@@ -129,6 +131,7 @@ impl<W: io::Write> CallConsensusRead<W> {
                 )?;
             }
             if record.is_unmapped() || record.is_mate_unmapped() {
+                unmark_record(&mut record)?;
                 self.bam_skipped_writer.write(&record)?;
                 continue;
             }
@@ -161,13 +164,15 @@ impl<W: io::Write> CallConsensusRead<W> {
                             //For right record save end position and duplicate group ID
                             let group_id_opt = match storage_entry {
                                 RecordStorage::PairedRecords {
-                                    r1_rec,
+                                    ref mut r1_rec,
                                     ref mut r2_rec,
                                 } => {
                                     let group_id = if cigar_has_softclips(r1_rec)
                                         || cigar_has_softclips(&record)
                                     {
+                                        unmark_record(r1_rec)?;
                                         self.bam_skipped_writer.write(r1_rec)?;
+                                        unmark_record(&mut record)?;
                                         self.bam_skipped_writer.write(&record)?;
                                         None
                                     } else {
@@ -189,7 +194,9 @@ impl<W: io::Write> CallConsensusRead<W> {
                                     let group_id = if cigar_has_softclips(rec)
                                         || cigar_has_softclips(&record)
                                     {
+                                        unmark_record(rec)?;
                                         self.bam_skipped_writer.write(rec)?;
+                                        unmark_record(&mut record)?;
                                         self.bam_skipped_writer.write(&record)?;
                                         None
                                     } else {
@@ -222,6 +229,7 @@ impl<W: io::Write> CallConsensusRead<W> {
                             if !record.is_paired() {
                                 //If right or single record save end position and duplicate group ID
                                 if cigar_has_softclips(&record) {
+                                    unmark_record(&mut record)?;
                                     self.bam_skipped_writer.write(&record)?;
                                 } else {
                                     duplicate_groups
@@ -265,6 +273,7 @@ impl<W: io::Write> CallConsensusRead<W> {
                         //Case: Left record
                         None => {
                             if !record.is_paired() || record.tid() != record.mtid() {
+                                unmark_record(&mut record)?;
                                 self.bam_skipped_writer.write(&record)?;
                             } else {
                                 record_storage.insert(
@@ -281,7 +290,7 @@ impl<W: io::Write> CallConsensusRead<W> {
                         }
                         //Case: Left record already stored
                         Some(_record_pair) => {
-                            let (rec_id, l_rec) = match record_storage
+                            let (rec_id, mut l_rec) = match record_storage
                                 .remove(&RecordId::Regular(record_name.to_owned()))
                                 .unwrap()
                             {
@@ -291,7 +300,9 @@ impl<W: io::Write> CallConsensusRead<W> {
                                 RecordStorage::SingleRecord { .. } => unreachable!(),
                             };
                             if cigar_has_softclips(&l_rec) || cigar_has_softclips(&record) {
+                                unmark_record(&mut l_rec)?;
                                 self.bam_skipped_writer.write(&l_rec)?;
+                                unmark_record(&mut record)?;
                                 self.bam_skipped_writer.write(&record)?;
                             } else {
                                 let alignment_vectors = calc_read_alignments(&l_rec, &record);
@@ -314,7 +325,9 @@ impl<W: io::Write> CallConsensusRead<W> {
                                         )?;
                                     }
                                     None => {
+                                        unmark_record(&mut l_rec)?;
                                         self.bam_skipped_writer.write(&l_rec)?;
+                                        unmark_record(&mut record)?;
                                         self.bam_skipped_writer.write(&record)?;
                                     }
                                 };
@@ -355,7 +368,7 @@ pub fn calc_consensus_complete_groups<'a, W: io::Write>(
     let group_ids = group_end_idx.cut_lower_group_ids(end_pos)?;
     for group_id in group_ids {
         let cigar_groups =
-            group_reads_by_cigar(duplicate_groups.remove(&group_id).unwrap(), record_storage)?;
+            group_reads(duplicate_groups.remove(&group_id).unwrap(), record_storage)?;
         for cigar_group in cigar_groups.values() {
             match cigar_group {
                 CigarGroup::PairedRecords {
@@ -388,23 +401,47 @@ pub fn calc_consensus_complete_groups<'a, W: io::Write>(
                             // If reads do not overlap or CIGAR in overlapping region differs R1 and R2 are handled sepperatly
                             if r1_recs.len() > 1 {
                                 let uuid = &Uuid::new_v4().to_hyphenated().to_string();
-                                fq1_writer.write_record(
-                                    &CalcNonOverlappingConsensus::new(
-                                        r1_recs, r1_seqids, uuid, read_ids,
-                                    )
-                                    .calc_consensus()
-                                    .0,
-                                )?;
-                                fq2_writer.write_record(
-                                    &CalcNonOverlappingConsensus::new(
-                                        r2_recs, r2_seqids, uuid, read_ids,
-                                    )
-                                    .calc_consensus()
-                                    .0,
-                                )?;
+                                match r1_recs[0].is_first_in_template() {
+                                    true => {
+                                        fq1_writer.write_record(
+                                            &CalcNonOverlappingConsensus::new(
+                                                r1_recs, r1_seqids, uuid, read_ids,
+                                            )
+                                            .calc_consensus()
+                                            .0,
+                                        )?;
+                                        fq2_writer.write_record(
+                                            &CalcNonOverlappingConsensus::new(
+                                                r2_recs, r2_seqids, uuid, read_ids,
+                                            )
+                                            .calc_consensus()
+                                            .0,
+                                        )?;
+                                    }
+                                    false => {
+                                        fq1_writer.write_record(
+                                            &CalcNonOverlappingConsensus::new(
+                                                r2_recs, r2_seqids, uuid, read_ids,
+                                            )
+                                            .calc_consensus()
+                                            .0,
+                                        )?;
+                                        fq2_writer.write_record(
+                                            &CalcNonOverlappingConsensus::new(
+                                                r1_recs, r1_seqids, uuid, read_ids,
+                                            )
+                                            .calc_consensus()
+                                            .0,
+                                        )?;
+                                    }
+                                }
                             } else {
-                                bam_skipped_writer.write(&r1_recs[0])?;
-                                bam_skipped_writer.write(&r2_recs[0])?;
+                                let mut r1_rec = r1_recs[0].clone();
+                                unmark_record(&mut r1_rec)?;
+                                bam_skipped_writer.write(&r1_rec)?;
+                                let mut r2_rec = r2_recs[0].clone();
+                                unmark_record(&mut r2_rec)?;
+                                bam_skipped_writer.write(&r2_rec)?;
                             }
                         }
                     };
@@ -419,7 +456,9 @@ pub fn calc_consensus_complete_groups<'a, W: io::Write>(
                         )?;
                     }
                     _ => {
-                        bam_skipped_writer.write(&recs[0])?;
+                        let mut rec = recs[0].clone();
+                        unmark_record(&mut rec)?;
+                        bam_skipped_writer.write(&rec)?;
                     }
                 },
             }
@@ -428,18 +467,18 @@ pub fn calc_consensus_complete_groups<'a, W: io::Write>(
     Ok(())
 }
 
-fn group_reads_by_cigar(
+// Final groups for consensus reads are created based on record and read orientations
+fn group_reads(
     record_ids: Vec<RecordId>,
     record_storage: &mut HashMap<RecordId, RecordStorage>,
-) -> Result<HashMap<Cigar, CigarGroup>> {
-    let mut cigar_groups: HashMap<Cigar, CigarGroup> = HashMap::new();
+) -> Result<HashMap<(Cigar, String), CigarGroup>> {
+    let mut final_groups: HashMap<(Cigar, String), CigarGroup> = HashMap::new();
     for rec_id in record_ids {
         let storage_entry = record_storage.remove(&rec_id).unwrap();
-        storage_entry.add_to_group(&mut cigar_groups)?;
+        storage_entry.add_to_group(&mut final_groups)?;
     }
-    Ok(cigar_groups)
+    Ok(final_groups)
 }
-
 fn calc_read_alignments(
     r1_rec: &bam::Record,
     r2_rec: &bam::Record,
@@ -498,7 +537,7 @@ fn calc_alignment_vectors(
     };
     let mut intersection_entry_passed = false;
     loop {
-        if r2_cigar == None {
+        if r2_cigar.is_none() {
             match r1_cigar {
                 None => break,
                 Some('M') | Some('X') | Some('=') | Some('D') | Some('N') => {
@@ -512,7 +551,7 @@ fn calc_alignment_vectors(
             if offset == 0 {
                 r2_cigar = r2_cigarstring.next();
             }
-        } else if r1_cigar == None {
+        } else if r1_cigar.is_none() {
             match_single_cigar(&r2_cigar, &mut r2_vec, &mut r1_vec);
             r2_cigar = r2_cigarstring.next();
         } else if r1_cigar != r2_cigar {
@@ -578,56 +617,67 @@ pub enum RecordStorage {
 }
 
 impl RecordStorage {
-    fn add_to_group(self, cigar_groups: &mut HashMap<Cigar, CigarGroup>) -> Result<()> {
-        let (r1_rec_entry, r1_rec_id, r2_rec_entry, r2_rec_id, cigar_tuple) = match self {
-            RecordStorage::PairedRecords { r1_rec, r2_rec } => {
-                let r1_rec_id = r1_rec.rec_id;
-                let r1_rec_entry = r1_rec.into_rec();
-                let r2_rec_unwrapped = r2_rec.unwrap();
-                let r2_rec_id = r2_rec_unwrapped.rec_id;
-                let r2_rec_entry = r2_rec_unwrapped.into_rec();
-                let cigar_tuple = Cigar::Tuple {
-                    r1_cigar: r1_rec_entry.raw_cigar().to_vec(),
-                    r2_cigar: r2_rec_entry.raw_cigar().to_vec(),
-                };
-                if !cigar_groups.contains_key(&cigar_tuple) {
-                    cigar_groups.insert(
-                        cigar_tuple.clone(),
-                        CigarGroup::PairedRecords {
+    fn add_to_group(self, final_groups: &mut HashMap<(Cigar, String), CigarGroup>) -> Result<()> {
+        let (r1_rec_entry, r1_rec_id, r2_rec_entry, r2_rec_id, cigar_tuple, orientation) =
+            match self {
+                RecordStorage::PairedRecords { r1_rec, r2_rec } => {
+                    let r1_rec_id = r1_rec.rec_id;
+                    let r1_rec_entry = r1_rec.into_rec();
+                    let r2_rec_unwrapped = r2_rec.unwrap();
+                    let r2_rec_id = r2_rec_unwrapped.rec_id;
+                    let r2_rec_entry = r2_rec_unwrapped.into_rec();
+                    let cigar_tuple = Cigar::Tuple {
+                        r1_cigar: r1_rec_entry.raw_cigar().to_vec(),
+                        r2_cigar: r2_rec_entry.raw_cigar().to_vec(),
+                    };
+                    let pair_orientation =
+                        r1_rec_entry.read_pair_orientation().as_ref().to_string();
+                    final_groups
+                        .entry((cigar_tuple.clone(), pair_orientation.to_string()))
+                        .or_insert_with(|| CigarGroup::PairedRecords {
                             r1_recs: Vec::new(),
                             r2_recs: Vec::new(),
                             r1_seqids: Vec::new(),
                             r2_seqids: Vec::new(),
-                        },
-                    );
+                        });
+                    (
+                        r1_rec_entry,
+                        r1_rec_id,
+                        Some(r2_rec_entry),
+                        Some(r2_rec_id),
+                        cigar_tuple,
+                        pair_orientation,
+                    )
                 }
-                (
-                    r1_rec_entry,
-                    r1_rec_id,
-                    Some(r2_rec_entry),
-                    Some(r2_rec_id),
-                    cigar_tuple,
-                )
-            }
-            RecordStorage::SingleRecord { rec } => {
-                let rec_id = rec.rec_id;
-                let rec_entry = rec.into_rec();
-                let cigar_single = Cigar::Single {
-                    cigar: rec_entry.raw_cigar().to_vec(),
-                };
-                if !cigar_groups.contains_key(&cigar_single) {
-                    cigar_groups.insert(
-                        cigar_single.clone(),
-                        CigarGroup::SingleRecords {
+                RecordStorage::SingleRecord { rec } => {
+                    let rec_id = rec.rec_id;
+                    let rec_entry = rec.into_rec();
+                    let cigar_single = Cigar::Single {
+                        cigar: rec_entry.raw_cigar().to_vec(),
+                    };
+                    let read_orientation = if rec_entry.is_reverse() {
+                        "reverse"
+                    } else {
+                        "forward"
+                    }
+                    .to_string();
+                    final_groups
+                        .entry((cigar_single.clone(), read_orientation.clone()))
+                        .or_insert_with(|| CigarGroup::SingleRecords {
                             recs: Vec::new(),
                             seqids: Vec::new(),
-                        },
-                    );
+                        });
+                    (
+                        rec_entry,
+                        rec_id,
+                        None,
+                        None,
+                        cigar_single,
+                        read_orientation,
+                    )
                 }
-                (rec_entry, rec_id, None, None, cigar_single)
-            }
-        };
-        match cigar_groups.get_mut(&cigar_tuple) {
+            };
+        match final_groups.get_mut(&(cigar_tuple, orientation)) {
             Some(CigarGroup::PairedRecords {
                 r1_recs,
                 r2_recs,
@@ -665,6 +715,12 @@ impl Deref for IndexedRecord {
     type Target = bam::Record;
     fn deref(&self) -> &bam::Record {
         &self.rec
+    }
+}
+
+impl DerefMut for IndexedRecord {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.rec
     }
 }