+++ src/fastq/collapse_reads_to_fragments/pipeline.rs
@@ -63,7 +63,6 @@ fn isize_pmf(value: f64, mean: f64, sd: f64) -> LogProb {
 #[derive(Debug)]
 struct FastqStorage {
     db: DB,
-    storage_dir: std::path::PathBuf,
 }
 
 impl FastqStorage {
@@ -74,8 +73,7 @@ impl FastqStorage {
         // in turn deleting the tempdir
         let storage_dir = tempdir()?.path().join("db");
         Ok(FastqStorage {
-            db: DB::open_default(storage_dir.clone())?,
-            storage_dir,
+            db: DB::open_default(storage_dir)?,
         })
     }
 
@@ -152,7 +150,7 @@ pub trait CallConsensusReads<'a, R: io::Read + io::BufRead + 'a, W: io::Write + 
         // prepare spinner for user feedback
         let pb = indicatif::ProgressBar::new_spinner();
         pb.set_style(spinner_style.clone());
-        pb.set_prefix(&"[1/2] Clustering input reads by UMI using starcode.".to_string());
+        pb.set_prefix("[1/2] Clustering input reads by UMI using starcode.");
 
         loop {
             // update spinner
@@ -199,7 +197,7 @@ pub trait CallConsensusReads<'a, R: io::Read + io::BufRead + 'a, W: io::Write + 
         let mut j = 0;
         let pb = indicatif::ProgressBar::new_spinner();
         pb.set_style(spinner_style);
-        pb.set_prefix(&"[1/2] Clustering input reads by UMI using starcode.".to_string());
+        pb.set_prefix("[1/2] Clustering input reads by UMI using starcode.");
         // read clusters identified by the first starcode run
         // the first run clustered by UMI, hence all reads in
         // the clusters handled here had similar UMIs
@@ -351,11 +349,11 @@ impl<'a, R: io::Read + io::BufRead, W: io::Write> CallConsensusReads<'a, R, W>
     }
 
     fn fq1_reader(&mut self) -> &mut fastq::Reader<R> {
-        &mut self.fq1_reader
+        self.fq1_reader
     }
 
     fn fq2_reader(&mut self) -> &mut fastq::Reader<R> {
-        &mut self.fq2_reader
+        self.fq2_reader
     }
 
     fn umi_len(&self) -> usize {
@@ -506,11 +504,11 @@ impl<'a, R: io::Read + io::BufRead, W: io::Write> CallConsensusReads<'a, R, W>
     }
 
     fn fq1_reader(&mut self) -> &mut fastq::Reader<R> {
-        &mut self.fq1_reader
+        self.fq1_reader
     }
 
     fn fq2_reader(&mut self) -> &mut fastq::Reader<R> {
-        &mut self.fq2_reader
+        self.fq2_reader
     }
 
     fn umi_len(&self) -> usize {